Audio Manger & Loading Screen
I have completed the loading screen for the main menu UI, so when you click on the tutorial then it will load the scene (currently we only have the sample scene). Also added in the audio  manager, however I could only add in background music as there is no shooting going on. Yes I could add in walking sound effects, however the motion script is a little basic for this. *Note I have not updated the credits section which should reference the audio sources (they are open source) will do this later on.